requirements.txt
fastapi>=0.110,<1.0
uvicorn>=0.23,<1.0
jinja2>=3.1,<4.0
httpx>=0.24,<1.0
pydantic>=2.0,<3.0
python-multipart>=0.0.6,<1.0
openpyxl>=3.1,<4.0
python-pptx>=0.6.23,<1.0
